refactor: create a constants file for tiptap editor extension
