+++ src/components/ui/drawers/tasks/task-description/TaskDescriptionRendererForCategory.tsx
@@ -1,5 +1,5 @@
 import { EditorContent, useEditor } from '@tiptap/react';
-import { editorExtensions } from '@/components/ui/drawers/categories/CategoryCreationDrawerContent';
+import { editorExtensions } from '@/constants/tiptap-editor';
 
 import { useFormContext } from 'react-hook-form';
 import { TCategoryCreationSchema } from '@/constants/schema';
